Rename store folder to redux to follow the Ducks Pattern
